requirements.txt
crewai
crewai_tools
langchain_openai
python-decouple
pydantic
langchain_community
slack-sdk
